📝 36993E task: record PR metadata
