+++ .agentplane/tasks/202605221110-36993E/pr/meta.json
@@ -5,9 +5,12 @@
   "head_sha": "87c0699ba34efd41c2aa64b85e6aeab18befdc63",
   "last_verified_at": "2026-05-22T11:28:58.936Z",
   "last_verified_sha": "87c0699ba34efd41c2aa64b85e6aeab18befdc63",
+  "pr_number": 4016,
+  "pr_url": "https://github.com/basilisk-labs/agentplane/pull/4016",
   "schema_version": 1,
+  "status": "OPEN",
   "task_id": "202605221110-36993E",
-  "updated_at": "2026-05-22T11:28:58.981Z",
+  "updated_at": "2026-05-22T11:29:19.735Z",
   "verify": {
     "status": "pass"
   }
